mvvm correction
