+++ app/src/main/java/ru/netology/nmedia/MainActivity.kt
@@ -4,7 +4,6 @@ import android.os.Bundle
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.observe
-import kotlinx.android.synthetic.main.activity_main.*
 import ru.netology.nmedia.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
@@ -20,6 +19,9 @@ class MainActivity : AppCompatActivity() {
                 author.text = post.author
                 published.text = post.published
                 text.text = post.content
+                textViewLikes.text = formatNumber(post.likes)
+                textViewShare.text = formatNumber(post.shares)
+                textViewViews.text = formatNumber(post.views)
                 imageViewLikes.setImageResource(
                     if (post.likedByMe) R.drawable.likes_liked
                     else R.drawable.likes
@@ -29,15 +31,45 @@ class MainActivity : AppCompatActivity() {
 
         binding.imageViewLikes.setOnClickListener {
             viewModel.like()
-            textViewLikes.text = viewModel.liked()
         }
 
         binding.imageViewShare.setOnClickListener {
-            textViewShare.text = viewModel.shared()
+            viewModel.shared()
         }
 
         binding.imageViewViews.setOnClickListener {
-            textViewViews.text = viewModel.viewed()
+            viewModel.viewed()
         }
+    }
+
+    private fun formatNumber(number: Int): String {
+        val suffixes = charArrayOf('k', 'm', 'g')
+        if (number < 1000) {
+            return java.lang.String.valueOf(number)
+        }
+
+        val string = java.lang.String.valueOf(number)
+
+        // разрядность числа
+        val magnitude = (string.length - 1) / 3
+
+        // количество цифр до суффикса
+        val digits = (string.length - 1) % 3 + 1
+
+        val value = CharArray(4)
+        for (i in 0 until digits) {
+            value[i] = string[i]
+        }
+        var valueLength = digits
+
+        // добавление точки и числа
+        if (digits == 1 && string[1] != '0') {
+            value[valueLength++] = '.'
+            value[valueLength++] = string[1]
+        }
+
+        // добавление суффикса
+        value[valueLength++] = suffixes[magnitude - 1]
+        return String(value)
     }
 }
