+++ app/src/main/java/ru/netology/nmedia/PostRepository.kt
@@ -5,7 +5,6 @@ import androidx.lifecycle.LiveData
 interface PostRepository {
     fun get(): LiveData<Post>
     fun like()
-    fun liked(): String
-    fun shared(): String
-    fun viewed(): String
+    fun shared()
+    fun viewed()
 }
