+++ app/src/main/java/ru/netology/nmedia/PostViewModel.kt
@@ -6,7 +6,6 @@ class PostViewModel : ViewModel() {
     private val repository: PostRepository = PostRepositoryInMemoryImpl()
     val data = repository.get()
     fun like() = repository.like()
-    fun liked() = repository.liked()
     fun shared() = repository.shared()
     fun viewed() = repository.viewed()
 }
